+++ core/body.h
@@ -14,10 +14,10 @@ enum class StellarType {
 
 class Body{
     public:
-        Body(std::string name_val, float mass_val, vector3d position_val, vector3d velocity_val, 
+        Body(std::string name_val, double mass_val, vector3d position_val, vector3d velocity_val, 
              double radius_val, StellarType type_val = StellarType::NONE, int binary_id_val = -1);
         std::string name;
-        float mass;
+        double mass;
         vector3d position;
         vector3d velocity;
         double radius;
